Extract shared logic to OcrProvider base class
